WIP 発言取得
