+++ components/type/chara/chara-default-message.ts
@@ -1,0 +1,6 @@
+interface CharaDefaultMessage {
+  join_message: string | null
+  first_day_message: string | null
+}
+
+export default CharaDefaultMessage
