+++ components/type/chara/chara-face.ts
@@ -1,0 +1,6 @@
+interface CharaFace {
+  type: string
+  image_url: string
+}
+
+export default CharaFace
