+++ components/type/chara/chara-name.ts
@@ -1,0 +1,6 @@
+interface CharaName {
+  name: string
+  short_name: string
+}
+
+export default CharaName
