+++ components/type/chara/chara-size.ts
@@ -1,0 +1,6 @@
+interface CharaSize {
+  width: number
+  height: number
+}
+
+export default CharaSize
