+++ components/type/chara/charachip.ts
@@ -1,0 +1,11 @@
+import Chara from '~/components/type/chara/chara'
+
+interface Charachip {
+  id: number
+  name: string
+  designer_id: number
+  description_utl: string
+  chara_list: Chara[]
+}
+
+export default Charachip
