+++ components/type/chara/charachips.ts
@@ -1,0 +1,7 @@
+import Charachip from '~/components/type/chara/charachip'
+
+interface Charachips {
+  charachip_list: Charachip[]
+}
+
+export default Charachips
