+++ components/type/skill/skill.ts
@@ -1,0 +1,6 @@
+interface Skill {
+  code: string
+  name: string
+}
+
+export default Skill
