+++ components/type/skill/skills.ts
@@ -1,0 +1,5 @@
+interface Skills {
+  hoge: string
+}
+
+export default Skills
