+++ components/type/village/village.ts
@@ -1,6 +1,7 @@
 import VillageStatus from '~/components/type/village/village-status'
 import VillageSettings from '~/components/type/village/setting/village-settings'
 import VillageParticipants from '~/components/type/village/participant/village-participants'
+import VillageDays from '~/components/type/village/village-days'
 
 interface Village {
   id: number
@@ -10,6 +11,7 @@ interface Village {
   setting: VillageSettings
   participant: VillageParticipants
   spectator: VillageParticipants
+  day: VillageDays
 }
 
 export default Village
